requirements.txt
streamlit
boto3
psycopg2-binary
pyodbc
langchain
langchain-community
langchain-openai
langchain-core
pgvector
python-dotenv
